Changed of Help activity's string IDs + minor edits to Splash Screen
Due to the last commit for the Main Menu branch also including chaning strings.xml for naming conventions, the string IDs used in the Help Activity needed to change
in order for the program to compile and run properly. Commit also includes minor edits to the (placeholder) title's location on the Splash Screen.