Added Storing of Different Card Amounts in PrefsManager
Redid the work for the previous commit in OptionsActivity,
Constants and PrefsManager after it was removed from an attempted
pull from master. Also added a way to store Different Card Amounts
(this depends on the Order, i.e: Order 2 has 7 different cards).
Setting defaults in the two Spinners is not yet complete.